reorganized the code into individual files,and created a go package on github
