+++ cxts/ContextKey.go
@@ -1,0 +1,3 @@
+package cxts
+
+type ContextKey string
