+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/mrpiggy97/golang-learning
+
+go 1.16
